+++ Broccolini/Parsing/IParserInput.cs
@@ -4,7 +4,13 @@ namespace Broccolini.Parsing;
 
 internal interface IParserInput
 {
+    int AvailableLength { get; }
+
     IniToken Peek(int lookAhead = 0);
 
+    IEnumerable<IniToken> PeekRange();
+
     IniToken Read();
+
+    ImmutableArray<IniToken> Read(IEnumerable<IniToken> peeked);
 }