+++ Broccolini/Syntax/IIniNodeVisitor.cs
@@ -7,9 +7,11 @@ public interface IIniNodeVisitor
 {
     void Visit(KeyValueIniNode keyValueNode);
 
-    void Visit(UnrecognizedIniNode triviaNode);
+    void Visit(UnrecognizedIniNode unrecognizedNode);
 
     void Visit(SectionIniNode sectionNode);
 
     void Visit(CommentIniNode commentNode);
+
+    void Visit(SectionHeaderIniNode sectionHeaderNode);
 }