+++ Broccolini/Syntax/IniToken.cs
@@ -111,6 +111,10 @@ public abstract record IniToken
 
     internal sealed record Epsilon : IniToken
     {
+        private Epsilon() { }
+
+        public static readonly Epsilon Instance = new();
+
         public override string ToString() => string.Empty;
 
         private protected override void InternalImplementorsOnly() { }